Mesh / Polygon2D conversion added
It is now possible to convert from Mesh Colliders to Polygon Collider 2Ds and the other way around! Both methods have some rather specific behavior that is important to understand. Collider2Dx has been set up to handle this automatically.

Mesh to PolygonCollider2D:
To convert from a mesh to a polygon, a camera renders the mesh (on layer 31, Unity's last render layer) to a texture. From this texture, a sprite is created. And this sprite generates a physics shape. The shapes points are then copied to the PolygonCollider2D, which gets offset by the same amount the mesh's center was offset from the mesh's transform.
What's important is that the size of the texture rendered to determines the precision of the generated collider, however maximum precision doesn't always yield the best results since points might get put so close together that they are discarded. Less precision also leads to better performance.

PolygonCollider2D to Mesh:
It is impossible to create a 3D mesh out of a PolygonCollider2D: it can only create flat meshes, and only when the PolygonCollider2D is active and enabled.
Since the created mesh is not owned by anyone, the user is responsible for its lifetime and not deleting the mesh will result in a memory leak. When creating a new mesh out of a PolygonCollider2D, it is possible to destroy the shared mesh. This is useful if you are constantly generating new shapes from the PolygonCollider2D and need to destroy every last iteration of the mesh.

Since both the ToPolygonCollider2D and ToMesh methods have parameters, the ToCollider methods have been altered to accept a ConversionSettings struct that can store these parameters.

Collider2Dx has been altered to pass ConversionSettings. Note that if a MeshCollider uses a mesh asset and the ConversionSettings specify that PolygonCollider2Ds should destroy a MeshColliders old mesh, an exception would be thrown because assets cannot be destroyed. To combat this, the Collider2Dx stores a copy of the MeshCollider's mesh if the ConversionSettings specify to destroy MeshCollider's old meshes.

Of Note:
When a MeshCollider is converted to a PolygonCollider2D and back many times over, the Mesh will gradually lose precision until it becomes a square.