General Fixes
ConversionSettings tolerance can now be any value (previously was clamped to 0 and 1).

Fixed a bug where MeshCollider's isTrigger could not be set if the MeshCollider was not convex. Additionally, if converting from a MeshCollider to a Polygon Collider, the Polygon Collider's isTrigger will also not be set if the MeshCollider is not convex to keep conversion consistent.

RuntimeComponentCopier has been deprecated in favor of custom copier functions that don't use (expensive) reflection.