Module2Dx Settings Update
Module2Dx no longer required an order variable. Instead the order of the list in the settings window is used.

Incidentally, the list is now reorderable for objects that aren't part of the Physics2Dx assembly, adding new Module2Dx classes to the conversion batching completely automatically! Checks are in place to ensure that internal Module2Dxes are not converted in a wrong order.

Since the index of the list is now used, Physics2Dx has been optimized to use arrays instead of KeyValuePairs to save on memory.

Old Editors have been removed.