Ignore Overlap
RigidbodyConverter now has a setting to specify that it needs to ignore colliders that are overlapping on conversion. It also doesn't use bounds anymore but uses Physics queries!

Old deprecated stuff has been removed. No more clutter!

Transform2Dx doesn't inherit from Module2Dx anymore, since the conversion method should be flexible (before, on, or after convert is a choice that depends on the project).

A Transform2DxConverter has been made that applies an on convert approach to set the Transform2Dx children active and inactive.