Rename to Dimension Converter
At the last minute, the new name Physics Converter was changed to the more encompassing Dimension Converter! Renaming has been done all around: Physics2Dx has been renamed to Dimension, Module2Dx is now Converter, Converter2Dx and Copier2Dx are now PhysicsConverter and PhysicsCopier... and more! The documentation still needs a thorough look, but the crux is that 2Dx will be removed from the naming.

RigidbodyConverter is documented now.

OnBeforeConvert and OnAfterConvert are 2 new behaviours that invoke a UnityEvent onBeforeConvert and onAfterConvert.