Outliner + Default Outliner
Different contour tracer settings can lead to different results. To scale this reliably across a project, an Outliner ScriptableObject has been created to store these settings.

The PhysicsConverter now contains a defaultOutliner field which will be used when no Outliner is specified in MeshToPolygon conversion. This defaultOutliner can be further specified in the settings.

The slimHierarchy property of the Settings has been removed. All DimensionConverter created objects will be shown always.

The documentation has been updated to include the new Outliner and the ContourTracer.