Add onRigidbodyConverted events
RigidbodyConverter will now throw an event when it has converted to using a Rigidbody. This event can later be used for Components like the JointConverters, which will need to know when the Rigidbody has been added. And if a script has a Rigidbody setter, it can be set using the inspector without any additional code!

All of the above applies to Rigidbody2D as well.