ConstantForceConverter added
Adds a ConstantForceConverter Component that can convert ConstantForces with ConstantForce2Ds.

Also adds the methods ToVector2D and ToVector to the utilities. These were previously used to convert a 3D offset to a 2D offset and back, but they also proved useful when working with the relativeForce of ConstantForce. Since these methods seem to have a higher underlying value, it has been decided to make them public and give with more generic naming.