Add SliderJoint2Dx
SliderJoint2Dx is the first prototype for Joint conversion. A whole structure needed to be implemented in order to maintain a Joint2Dxs connection to a connected body through conversion.
SliderJoint2Dx, like all the other 2Dx classes, allows to convert between a SliderJoint and a SliderJoint2D. Though not finalized, it paves the way for more Joint conversions to be added in the future.